Move log depot manual files to separate module
